Add gdal_wgs84_to_egm08_working.sh script.
This script tests the current environment to verify than GDAL is properly
applying the vertical datum transformation between WGS84 and EGM08 based
on an (arbitrarily selected) sample pixel from the pgc-opendata-dems S3
bucket.

If the grid transform files are not available to PROJ, PROJ (and thus GDAL)
will silently apply the null transform instead.

The first test checks that the source data is as expected,
the second test checks the vdatum transform works.

Test successful:
	Testing if Proj Datum Grids are installed correctly...
	Test WGS84 elevation...
	Expected: -53.859375
	Got: -53.859375

	Test EGM08 elevation...
	Expected: 1.05549550056458
	Got: 1.05549550056458

Test failed (and exits 1):
	Testing if Proj Datum Grids are installed correctly...
	Test WGS84 elevation...
	Expected: -53.859375
	Got: -53.859375

	Test EGM08 elevation...
	Expected: 1.05549550056458
	Got: -53.859375
	Failed